+++ CHANGELOG.md
@@ -14,9 +14,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 - Simplified error message for an invalid range.
+- Simplified output of `--help-pattern` and `--help-filters` flags.
 - Output of `-h, --help` flag is organized into sections.
 - Output of `-h, --help` flag uses more colors in descriptions.
-- Output of `--help-pattern` and `--help-filters` flags matches online documentation.
 
 ### Fixed